Migrate some scheduled queries endpoints
